Remove the introspect script from image_manager directory
